requirements.txt
amqp==2.1.4
anyjson==0.3.3
autoflake==0.7
awesome-slugify==1.6.2
beautifulsoup4==4.2.1
billiard==3.5.0.2
boto==2.38.0
# celery==3.1.18
certifi==2017.7.27.1
chardet==3.0.4
coverage==4.4.1
decorator==4.1.2
diff-match-patch==20121119
dj-database-url==0.4.2
Django==1.8.7
django-activity-stream==0.6.1
django-appconf==0.5
django-autocomplete-light==2.3.3
django-autofixture==0.12.1
django-autoslug==1.9.3
django-bootstrap-form==3.3
django-bootstrap3-datetimepicker==2.2.3
django-braces==1.10.0
# django-celery==3.1.16
django-cors-headers==2.1.0
django-downloadview==1.2
django-extensions==1.6.1
django-floppyforms==1.7.0
django-forms-bootstrap==3.0.1
django-friendly-tag-loader==1.2.1
django-geoexplorer==4.0.5
django-geonode-client==0.0.15
django-guardian==1.4.6
django-haystack==2.4.1
django-import-export==0.5.1
django-jsonfield==0.9.16
django-jsonfield-compat==0.4.4
django-leaflet==0.13.7
django-modeltranslation==0.12
django-mptt==0.8.6
django-multi-email-field==0.5.1
django-nose==1.4.4
django-notify-x==0.1.6
django-oauth-toolkit==0.12.0
django-pagination==1.0.7
django-polymorphic==0.9.2
django-reversion==2.0.10
django-storages==1.1.8
django-suit==0.2.21
django-taggit==0.21.0
django-tastypie==0.12.2
django-treebeard==3.0
django-utils==0.0.2
elasticsearch==2.4.0
et-xmlfile==1.0.1
factory-boy==2.9.2
Faker==0.8.4
flake8==2.3.0
funcsigs==1.0.2
geolinks==0.2.0
geonode-agon-ratings==0.3.5
geonode-announcements==1.0.8
geonode-arcrest==10.2
geonode-avatar==2.1.6
geonode-dialogos==0.7
geonode-notification==1.1.1
geonode-user-accounts==1.0.15
# geonode-user-messages==0.1.5
gisdata==0.5.4
gsconfig==1.0.6
gsimporter==1.0.0
httplib2==0.9.2
idna==2.6
ipaddress==1.0.18
jdcal==1.3
kombu==4.1.0
lxml==3.6.4
mccabe==0.5.2
mock==2.0.0
MultipartPostHandler==0.1.0
nose==1.3.7
numpy==1.13.1
oauthlib==2.0.1
odfpy==1.3.5
openpyxl==2.4.8
OWSLib==0.11.0
Paver==1.2.4
pbr==3.1.1
#pdb==0.1
pep8==1.6.2
Pillow==3.3.1
Pinax==0.9a2
pinax-notifications==4.0.0
pinax-theme-bootstrap==3.0a11
pinax-theme-bootstrap-account==1.0b2
psutil==5.4.0
psycopg2==2.7.3.1
pycsw==2.0.2
pyelasticsearch==0.6.1
pyflakes==1.2.3
PyJWT==1.5.3
pyproj==1.9.5.1
pyshp==1.2.3
python-dateutil==2.5.3
python-gnupg==0.4.1
python-mimeparse==1.5.2
python-openid==2.2.5
python-social-auth==0.2.19
pytz==2016.6.1
PyYAML==3.11
redis==2.10.5
regex==2016.7.21
requests==2.11.1
requests-oauthlib==0.8.0
Shapely==1.5.13
simplejson==3.11.1
six==1.10.0
social-auth-core==1.5.0
tablib==0.12.1
transifex-client==0.10
ua-parser==0.7.3
unicodecsv==0.14.1
Unidecode==0.4.19
urllib3==1.22
user-agents==1.1.0
vine==1.1.4
xlrd==1.1.0
xlwt==1.3.0
xmljson==0.1.9
xmltodict==0.9.2

django-filter==1.0.4
djangorestframework==3.6
djangorestframework-gis==0.11.2
django-db-logger==0.1.6

celery==4.2
markdown==2.6.9

notebook==4.4.1

# Do not remove below lines
# pip install -e git+https://github.com/celery/billiard.git#egg=billiard
# pip install -e git+https://github.com/celery/kombu.git#egg=kombu
